refactor: rename getShareOffer file old useShareOffer -> getShareOffer new getShareOffer -> getShareOfferBest
