+++ src/features/bookOffer/pages/BookingConfirmation.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useOffer } from 'features/offer/api/useOffer'
-import { getShareOffer } from 'features/share/helpers/getShareOffer'
+import { getShareOffer } from 'features/share/helpers/getShareOfferBest'
 import { WebShareModal } from 'features/share/pages/WebShareModalBest'
 import { analytics } from 'libs/analytics'
 import { useShowReview } from 'libs/hooks/useShowReview'
