+++ src/features/offer/components/OfferHeader/OfferHeader.tsx
@@ -5,7 +5,7 @@ import { useTheme } from 'styled-components/native'
 import { OfferResponse } from 'api/gen'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
-import { getShareOffer } from 'features/share/helpers/getShareOffer'
+import { getShareOffer } from 'features/share/helpers/getShareOfferBest'
 import { WebShareModal } from 'features/share/pages/WebShareModalBest'
 import { analytics } from 'libs/analytics'
 import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
